requirements.txt
asttokens==3.0.0
attrs==25.3.0
Authlib==1.5.2
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.1.0
Flask-Login==0.6.3
idna==3.10
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pip==23.2.1
pipreqs==0.5.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
pywin32==310
pyzmq==26.4.0
ratelimit==2.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
setuptools==65.5.0
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
yarg==0.1.9
